+++ src/main/resources/templates/signin.html
@@ -3,25 +3,118 @@
 <head>
     <meta charset="UTF-8">
     <title>로그인</title>
+    <style>
+        /* 전체 배경 및 폰트 설정 */
+        body {
+            margin: 0;
+            padding: 0;
+            background: linear-gradient(135deg, #f0f0f0, #e0e0e0);
+            font-family: -apple-system, BlinkMacSystemFont, "Helvetica Neue", "Apple SD Gothic Neo", "Malgun Gothic", sans-serif;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+        }
+
+        .logo {
+            display: block;
+            margin: 0 auto 1.5rem;
+            width: 90%;
+            margin-top: 100px;
+            margin-bottom: 30px;
+        }
+
+        /* 회원가입 컨테이너 스타일 */
+        .signup-container {
+            background-color: #fff;
+            border-radius: 12px;
+            box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+            width: 90%;
+            max-width: 400px;
+            padding: 2rem;
+            height: 100vh;
+            display: flex;
+            flex-direction: column;
+        }
+
+        h2 {
+            text-align: center;
+            font-size: 1.8rem;
+            margin-bottom: 1.5rem;
+            color: #333;
+        }
+
+        form {
+            display: flex;
+            flex-direction: column;
+        }
+
+        .form-group {
+            margin-bottom: 1rem;
+        }
+
+        label {
+            font-size: 0.9rem;
+            margin-bottom: 0.5rem;
+            color: #555;
+        }
+
+        input[type="text"],
+        input[type="email"],
+        input[type="password"],
+        input[type="tel"],
+        input[type="url"] {
+            padding: 12px;
+            border: 1px solid #ccc;
+            border-radius: 8px;
+            font-size: 1rem;
+            width: 100%;
+            box-sizing: border-box;
+            transition: border-color 0.3s ease;
+        }
+
+        input:focus {
+            border-color: #007aff;
+            outline: none;
+        }
+
+        button {
+            padding: 12px;
+            border: none;
+            border-radius: 8px;
+            background-color: #007aff;
+            color: #fff;
+            font-size: 1rem;
+            font-weight: bold;
+            cursor: pointer;
+            transition: background-color 0.3s ease;
+            width: 100%;
+        }
+
+        button:hover {
+            background-color: #005bb5;
+        }
+    </style>
 </head>
 <body>
-<div class="container">
-
-    <div class="py-4 text-center">
-        <h2>로그인</h2>
-    </div>
+<div class="signup-container">
+    <img th:src="@{/img/levelup-logo.png}" alt="Levelup 로고" class="logo">
     <form method="post" th:action="@{/v2/signin}">
-        <div>
-            <label for="email">이메일:</label>
-            <input id="email" name="email" required th:value="${email}" type="text">
+
+        <div class="form-group">
+            <input id="email" name="email" required th:value="${email}" type="text" placeholder="이메일 입력">
         </div>
-        <div>
-            <label for="password">비밀번호:</label>
-            <input id="password" name="nickName" required th:value="${password}" type="password">
+        <div class="form-group">
+            <input id="password" name="nickName" required th:value="${password}" type="password" placeholder="비밀번호 입력">
         </div>
-        <button type="submit">로그인</button>
+        <div class="form-group">
+            <button type="submit">로그인</button>
+        </div>
     </form>
-    <input id="autoLogin" name="autoLogin" onchange="AutoLogin()" type="checkbox">자동로그인</input>
+    <label for="autoLogin">
+        <input id="autoLogin" name="autoLogin" onchange="AutoLogin()" type="checkbox">
+        자동로그인
+    </label>
     <a href="/v2/signin/oauth2/authorization/naver">
         <img alt="네이버 로그인 버튼"
              height="50"
